refactor(components): split LaunchPanel into launch-panel subdirectory
Splits the 1040-line LaunchPanel.tsx into a thin parent plus
components/launch-panel/ with types, helpers, focusTrap, PreviewModal,
LaunchPanelFeedback, LaunchPanelControls, and LaunchPanelVersionStatus
subcomponents.

Default export, props shape, BEM class names, and aria attributes all
preserved. scripts/lint.sh and npm run build both green. Manual browser
smoke deferred to reviewer.

Part of #290
