+++ src/crosshook-native/src/components/LaunchPanel.tsx
@@ -1,5 +1,4 @@
-import { type KeyboardEvent, type MouseEvent, type ReactNode, useEffect, useId, useRef, useState } from 'react';
-import { createPortal } from 'react-dom';
+import { useEffect, useId, useState } from 'react';
 import { useLaunchStateContext } from '../context/LaunchStateContext';
 import { useProfileHealthContext } from '../context/ProfileHealthContext';
 import {
@@ -7,651 +6,17 @@ import {
   useAcknowledgeVersionChange,
 } from '../hooks/useAcknowledgeVersionChange';
 import { usePreviewState } from '../hooks/usePreviewState';
-import type {
-  EnvVarSource,
-  LaunchMethod,
-  LaunchPreview,
-  LaunchRequest,
-  LaunchValidationSeverity,
-  PatternMatch,
-  PreviewEnvVar,
-} from '../types';
 import { LaunchPhase } from '../types';
-import type { GameProfile } from '../types/profile';
 import { copyToClipboard } from '../utils/clipboard';
-import { LAUNCH_PANEL_ACTION_BUTTON_STYLE } from '../utils/launchPanelActionButtonStyle';
-import { sortIssuesBySeverity } from '../utils/mapValidationToNode';
-import { InfoCircleIcon } from './icons/SidebarIcons';
 import { LaunchPipeline } from './LaunchPipeline';
-import { CollapsibleSection } from './ui/CollapsibleSection';
+import { LaunchPanelControls } from './launch-panel/LaunchPanelControls';
+import { LaunchPanelFeedback } from './launch-panel/LaunchPanelFeedback';
+import { LaunchPanelVersionStatus } from './launch-panel/LaunchPanelVersionStatus';
+import { PreviewModal } from './launch-panel/PreviewModal';
+import type { LaunchPanelProps } from './launch-panel/types';
 import '../styles/preview.css';
 
-const UMU_DATABASE_MISSING_HINT =
-  'umu has no known entry for this app id in the current umu database. The database only tracks titles needing protonfixes — most titles work fine without an entry.';
-
-/* ───────── Focus-trap helpers (mirrors ProfileReviewModal) ───────── */
-
-const FOCUSABLE_SELECTOR = [
-  'a[href]',
-  'button:not([disabled])',
-  'input:not([disabled]):not([type="hidden"])',
-  'select:not([disabled])',
-  'textarea:not([disabled])',
-  '[tabindex]:not([tabindex="-1"])',
-  '[contenteditable="true"]',
-].join(', ');
-
-function getFocusableElements(container: HTMLElement): HTMLElement[] {
-  return Array.from(container.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)).filter(
-    (el) => !el.hasAttribute('disabled') && el.tabIndex >= 0 && el.getClientRects().length > 0
-  );
-}
-
-function focusElement(element: HTMLElement | null): boolean {
-  if (!element) return false;
-  element.focus({ preventScroll: true });
-  return document.activeElement === element;
-}
-
-/* ───────── Preview data helpers ───────── */
-
-function severityIcon(severity: LaunchValidationSeverity): string {
-  switch (severity) {
-    case 'fatal':
-      return '\u2717';
-    case 'warning':
-      return '!';
-    default:
-      return '\u2713';
-  }
-}
-
-function sortPatternMatchesBySeverity(matches: PatternMatch[]): PatternMatch[] {
-  const order: Record<LaunchValidationSeverity, number> = { fatal: 0, warning: 1, info: 2 };
-  return [...matches].sort((a, b) => order[a.severity] - order[b.severity]);
-}
-
-function sourceLabel(source: EnvVarSource): string {
-  switch (source) {
-    case 'proton_runtime':
-      return 'Proton Runtime';
-    case 'launch_optimization':
-      return 'Launch Optimization';
-    case 'host':
-      return 'Host';
-    case 'steam_proton':
-      return 'Steam Proton';
-    case 'profile_custom':
-      return 'Profile custom';
-  }
-}
-
-function groupEnvBySource(vars: PreviewEnvVar[]): [string, PreviewEnvVar[]][] {
-  const groups = new Map<string, PreviewEnvVar[]>();
-  for (const v of vars) {
-    const label = sourceLabel(v.source);
-    const list = groups.get(label);
-    if (list) {
-      list.push(v);
-    } else {
-      groups.set(label, [v]);
-    }
-  }
-  return Array.from(groups.entries());
-}
-
-function methodLabel(method: string): string {
-  switch (method) {
-    case 'steam_applaunch':
-      return 'Steam Launch';
-    case 'proton_run':
-      return 'Proton Launch';
-    case 'native':
-      return 'Native Launch';
-    default:
-      return method;
-  }
-}
-
-function isStale(generatedAt: string): boolean {
-  const generatedTime = new Date(generatedAt).getTime();
-  if (!Number.isFinite(generatedTime)) {
-    return true;
-  }
-
-  return Date.now() - generatedTime > 60_000;
-}
-
-function buildSummaryParts(preview: LaunchPreview): ReactNode[] {
-  const envCount = preview.environment?.length ?? 0;
-  const wrapperCount = preview.wrappers?.length ?? 0;
-  const fatalCount = preview.validation.issues.filter((i) => i.severity === 'fatal').length;
-  const warningCount = preview.validation.issues.filter((i) => i.severity === 'warning').length;
-  const passedCount = preview.validation.issues.filter((i) => i.severity === 'info').length;
-
-  const parts: ReactNode[] = [
-    <span key="env" className="crosshook-preview-modal__count--success">
-      {envCount} env vars
-    </span>,
-    <span key="wrap" className="crosshook-preview-modal__count--success">
-      {wrapperCount} {wrapperCount === 1 ? 'wrapper' : 'wrappers'}
-    </span>,
-  ];
-
-  if (passedCount > 0) {
-    parts.push(
-      <span key="pass" className="crosshook-preview-modal__count--success">
-        {passedCount} passed
-      </span>
-    );
-  }
-  if (warningCount > 0) {
-    parts.push(
-      <span key="warn" className="crosshook-preview-modal__count--warning">
-        {warningCount} {warningCount === 1 ? 'warning' : 'warnings'}
-      </span>
-    );
-  }
-  if (fatalCount > 0) {
-    parts.push(
-      <span key="err" className="crosshook-preview-modal__count--danger">
-        {fatalCount} {fatalCount === 1 ? 'error' : 'errors'}
-      </span>
-    );
-  }
-  if (preview.validation.issues.length === 0) {
-    parts.push(
-      <span key="ok" className="crosshook-preview-modal__count--success">
-        all checks passed
-      </span>
-    );
-  }
-
-  return parts;
-}
-
-/* ───────── PreviewModal component ───────── */
-
-interface PreviewModalProps {
-  preview: LaunchPreview;
-  profileId: string;
-  onClose: () => void;
-  onLaunch: () => void;
-}
-
-function PreviewModal({ preview, profileId, onClose, onLaunch }: PreviewModalProps) {
-  const portalHostRef = useRef<HTMLElement | null>(null);
-  const surfaceRef = useRef<HTMLDivElement | null>(null);
-  const headingRef = useRef<HTMLHeadingElement | null>(null);
-  const previouslyFocusedRef = useRef<HTMLElement | null>(null);
-  const bodyStyleRef = useRef('');
-  const hiddenNodesRef = useRef<Array<{ element: HTMLElement; inert: boolean; ariaHidden: string | null }>>([]);
-  const titleId = useId();
-  const [isMounted, setIsMounted] = useState(false);
-  const [copyLabel, setCopyLabel] = useState('Copy Preview');
-
-  useEffect(() => {
-    const host = document.createElement('div');
-    host.className = 'crosshook-modal-portal';
-    portalHostRef.current = host;
-    document.body.appendChild(host);
-    setIsMounted(true);
-
-    return () => {
-      host.remove();
-      portalHostRef.current = null;
-      setIsMounted(false);
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!isMounted) return;
-
-    const { body } = document;
-    const portalHost = portalHostRef.current;
-    if (!portalHost) return;
-
-    previouslyFocusedRef.current = document.activeElement instanceof HTMLElement ? document.activeElement : null;
-
-    bodyStyleRef.current = body.style.overflow;
-    body.style.overflow = 'hidden';
-    body.classList.add('crosshook-modal-open');
-
-    hiddenNodesRef.current = Array.from(body.children)
-      .filter((child): child is HTMLElement => child instanceof HTMLElement && child !== portalHost)
-      .map((element) => {
-        const inertState = (element as HTMLElement & { inert?: boolean }).inert ?? false;
-        const ariaHidden = element.getAttribute('aria-hidden');
-        (element as HTMLElement & { inert?: boolean }).inert = true;
-        element.setAttribute('aria-hidden', 'true');
-        return { element, inert: inertState, ariaHidden };
-      });
-
-    const frame = window.requestAnimationFrame(() => {
-      if (focusElement(headingRef.current)) return;
-      const focusable = surfaceRef.current ? getFocusableElements(surfaceRef.current) : [];
-      if (focusable.length > 0) focusElement(focusable[0]);
-    });
-
-    return () => {
-      window.cancelAnimationFrame(frame);
-      body.style.overflow = bodyStyleRef.current;
-      body.classList.remove('crosshook-modal-open');
-
-      for (const { element, inert, ariaHidden } of hiddenNodesRef.current) {
-        (element as HTMLElement & { inert?: boolean }).inert = inert;
-        if (ariaHidden === null) {
-          element.removeAttribute('aria-hidden');
-        } else {
-          element.setAttribute('aria-hidden', ariaHidden);
-        }
-      }
-      hiddenNodesRef.current = [];
-
-      const restoreTarget = previouslyFocusedRef.current;
-      if (restoreTarget?.isConnected) {
-        focusElement(restoreTarget);
-      }
-      previouslyFocusedRef.current = null;
-    };
-  }, [isMounted]);
-
-  function handleKeyDown(event: KeyboardEvent<HTMLDivElement>) {
-    if (event.key === 'Escape') {
-      event.stopPropagation();
-      event.preventDefault();
-      onClose();
-      return;
-    }
-
-    if (event.key !== 'Tab') return;
-
-    const container = surfaceRef.current;
-    if (!container) return;
-
-    const focusable = getFocusableElements(container);
-    if (focusable.length === 0) {
-      event.preventDefault();
-      return;
-    }
-
-    const currentIndex = focusable.indexOf(document.activeElement as HTMLElement);
-    const lastIndex = focusable.length - 1;
-
-    if (event.shiftKey) {
-      if (currentIndex <= 0) {
-        event.preventDefault();
-        focusElement(focusable[lastIndex]);
-      }
-      return;
-    }
-
-    if (currentIndex === -1 || currentIndex === lastIndex) {
-      event.preventDefault();
-      focusElement(focusable[0]);
-    }
-  }
-
-  function handleBackdropMouseDown(event: MouseEvent<HTMLDivElement>) {
-    if (event.target !== event.currentTarget) return;
-    onClose();
-  }
-
-  async function handleCopy() {
-    try {
-      await copyToClipboard(preview.display_text);
-      setCopyLabel('Copied');
-      window.setTimeout(() => setCopyLabel('Copy Preview'), 2000);
-    } catch {
-      setCopyLabel('Copy failed');
-      window.setTimeout(() => setCopyLabel('Copy Preview'), 2000);
-    }
-  }
-
-  if (!isMounted || !portalHostRef.current) return null;
-
-  const sortedIssues = sortIssuesBySeverity(preview.validation.issues);
-  const summaryParts = buildSummaryParts(preview);
-  const stale = isStale(preview.generated_at);
-  const isNative = preview.resolved_method === 'native';
-  const isSteam = preview.resolved_method === 'steam_applaunch';
-  const envCount = preview.environment?.length ?? 0;
-  const previewIsReady = preview.validation.issues.length === 0;
-  const generatedTime = new Date(preview.generated_at);
-  const generatedTimeLabel = Number.isFinite(generatedTime.getTime())
-    ? generatedTime.toLocaleTimeString()
-    : 'time unavailable';
-
-  return createPortal(
-    <div className="crosshook-modal" role="presentation">
-      <div className="crosshook-modal__backdrop" aria-hidden="true" onMouseDown={handleBackdropMouseDown} />
-      <div
-        ref={surfaceRef}
-        className="crosshook-modal__surface crosshook-panel crosshook-focus-scope"
-        role="dialog"
-        aria-modal="true"
-        aria-labelledby={titleId}
-        data-crosshook-focus-root="modal"
-        onKeyDown={handleKeyDown}
-      >
-        {/* Header */}
-        <header className="crosshook-modal__header">
-          <div className="crosshook-modal__heading-block">
-            <div className="crosshook-heading-eyebrow">Launch preview</div>
-            <h2 ref={headingRef} id={titleId} className="crosshook-modal__title" tabIndex={-1}>
-              {profileId}
-            </h2>
-          </div>
-          <div className="crosshook-modal__header-actions">
-            <span
-              className={[
-                'crosshook-modal__status-chip',
-                previewIsReady ? 'crosshook-modal__status-chip--success' : 'crosshook-modal__status-chip--danger',
-              ].join(' ')}
-            >
-              {previewIsReady ? 'Ready to launch' : 'Cannot launch'}
-            </span>
-          </div>
-        </header>
-
-        {/* Summary Banner */}
-        <div className="crosshook-preview-modal__summary-banner">
-          <div className="crosshook-preview-modal__summary-fields">
-            <div className="crosshook-preview-modal__summary-field">
-              <div className="crosshook-preview-modal__summary-label">Method</div>
-              <div className="crosshook-preview-modal__summary-value">{methodLabel(preview.resolved_method)}</div>
-            </div>
-            <div className="crosshook-preview-modal__summary-field">
-              <div className="crosshook-preview-modal__summary-label">Game Executable</div>
-              <div className="crosshook-preview-modal__summary-value crosshook-preview-modal__summary-value--mono">
-                {preview.game_executable_name || preview.game_executable}
-              </div>
-            </div>
-            <div className="crosshook-preview-modal__summary-field">
-              <div className="crosshook-preview-modal__summary-label">Working Directory</div>
-              <div className="crosshook-preview-modal__summary-value crosshook-preview-modal__summary-value--mono">
-                {preview.working_directory || 'Not set'}
-              </div>
-            </div>
-          </div>
-          <p className="crosshook-preview-modal__plan-line">
-            Preview:{' '}
-            {summaryParts.reduce<ReactNode[]>((acc, part, i) => {
-              if (i > 0) acc.push(', ');
-              acc.push(part);
-              return acc;
-            }, [])}
-          </p>
-        </div>
-
-        {/* Body with collapsible sections */}
-        <div className="crosshook-modal__body">
-          <div className="crosshook-preview-modal__sections">
-            {/* Validation Results */}
-            <CollapsibleSection
-              title="Validation Results"
-              defaultOpen
-              meta={
-                <span style={{ fontSize: '0.82rem' }}>
-                  {preview.validation.issues.length} {preview.validation.issues.length === 1 ? 'check' : 'checks'}
-                </span>
-              }
-            >
-              {sortedIssues.length > 0 ? (
-                <ul className="crosshook-preview-modal__validation-list">
-                  {sortedIssues.map((issue) => (
-                    <li
-                      key={`${issue.severity}-${issue.code ?? 'nocode'}-${issue.message}-${issue.help}-${issue.trainer_hash_stored ?? ''}-${issue.trainer_hash_current ?? ''}-${issue.trainer_sha256_community ?? ''}`}
-                      className="crosshook-preview-modal__validation-item"
-                      data-severity={issue.severity}
-                    >
-                      <span
-                        className="crosshook-preview-modal__validation-icon"
-                        data-severity={issue.severity}
-                        aria-hidden="true"
-                      >
-                        {severityIcon(issue.severity)}
-                      </span>
-                      <div className="crosshook-preview-modal__validation-content">
-                        <div className="crosshook-preview-modal__validation-message">{issue.message}</div>
-                        {issue.help ? (
-                          <div className="crosshook-preview-modal__validation-help">{issue.help}</div>
-                        ) : null}
-                      </div>
-                    </li>
-                  ))}
-                </ul>
-              ) : (
-                <p className="crosshook-preview-modal__empty">All validation checks passed.</p>
-              )}
-            </CollapsibleSection>
-
-            {/* Command Chain */}
-            <CollapsibleSection title="Command Chain" defaultOpen>
-              {preview.effective_command ? (
-                <pre className="crosshook-preview-modal__command-block">{preview.effective_command}</pre>
-              ) : (
-                <p className="crosshook-preview-modal__empty">No command resolved.</p>
-              )}
-              {preview.umu_decision
-                ? (() => {
-                    const umuChipModifier =
-                      preview.umu_decision.will_use_umu && preview.umu_decision.csv_coverage === 'missing'
-                        ? 'crosshook-preview-modal__umu-decision--info'
-                        : preview.umu_decision.will_use_umu
-                          ? 'crosshook-preview-modal__umu-decision--umu'
-                          : 'crosshook-preview-modal__umu-decision--proton';
-                    return (
-                      <div className={['crosshook-preview-modal__umu-decision', umuChipModifier].join(' ')}>
-                        <div>
-                          <strong>umu decision:</strong>{' '}
-                          {preview.umu_decision.will_use_umu ? 'using umu-run' : 'using direct Proton'}
-                        </div>
-                        <div className="crosshook-muted">
-                          requested preference: <code>{preview.umu_decision.requested_preference}</code>
-                          {' · '}
-                          umu-run on PATH:{' '}
-                          <code>{preview.umu_decision.umu_run_path_on_backend_path ?? 'not found'}</code>
-                        </div>
-                        <div className="crosshook-muted crosshook-preview-modal__umu-decision-reason">
-                          {preview.umu_decision.reason}
-                        </div>
-                        <div className="crosshook-muted" style={{ marginTop: 4 }}>
-                          umu protonfix coverage: <code>{preview.umu_decision.csv_coverage}</code>
-                        </div>
-                        {preview.umu_decision.will_use_umu && preview.umu_decision.csv_coverage === 'missing' ? (
-                          <div className="crosshook-preview-modal__umu-decision-info">
-                            <InfoCircleIcon
-                              className="crosshook-preview-modal__umu-decision-info-icon"
-                              width={14}
-                              height={14}
-                              aria-hidden
-                            />
-                            <span>{UMU_DATABASE_MISSING_HINT}</span>
-                          </div>
-                        ) : null}
-                      </div>
-                    );
-                  })()
-                : null}
-              {isSteam && preview.steam_launch_options ? (
-                <>
-                  <div className="crosshook-preview-modal__sub-label">Steam Launch Options</div>
-                  <pre className="crosshook-preview-modal__command-block">{preview.steam_launch_options}</pre>
-                </>
-              ) : null}
-              {preview.directives_error ? (
-                <div className="crosshook-preview-modal__directives-error">{preview.directives_error}</div>
-              ) : null}
-            </CollapsibleSection>
-
-            {/* Environment Variables */}
-            <CollapsibleSection
-              title="Environment Variables"
-              defaultOpen={false}
-              meta={<span style={{ fontSize: '0.82rem' }}>{envCount} vars</span>}
-            >
-              {preview.environment && preview.environment.length > 0 ? (
-                groupEnvBySource(preview.environment).map(([group, vars]) => (
-                  <div key={group} className="crosshook-preview-modal__env-group">
-                    <div className="crosshook-preview-modal__env-group-title">{group}</div>
-                    <table className="crosshook-preview-modal__env-table">
-                      <tbody>
-                        {vars.map((v) => (
-                          <tr key={v.key}>
-                            <td>{v.key}</td>
-                            <td>{v.value}</td>
-                          </tr>
-                        ))}
-                      </tbody>
-                    </table>
-                  </div>
-                ))
-              ) : (
-                <p className="crosshook-preview-modal__empty">No environment variables resolved.</p>
-              )}
-              {preview.cleared_variables.length > 0 ? (
-                <div className="crosshook-preview-modal__cleared-vars">
-                  <div className="crosshook-preview-modal__sub-label">Cleared Variables</div>
-                  <ul className="crosshook-preview-modal__cleared-list">
-                    {preview.cleared_variables.map((v) => (
-                      <li key={v} className="crosshook-preview-modal__cleared-item">
-                        {v}
-                      </li>
-                    ))}
-                  </ul>
-                </div>
-              ) : null}
-            </CollapsibleSection>
-
-            {/* Proton / Runtime Setup — hidden for native method (BR-8) */}
-            {!isNative ? (
-              <CollapsibleSection title="Proton / Runtime Setup" defaultOpen={false}>
-                {preview.proton_setup ? (
-                  <div className="crosshook-preview-modal__field-list">
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Wine Prefix</div>
-                      <div className="crosshook-preview-modal__field-value">
-                        {preview.proton_setup.wine_prefix_path}
-                      </div>
-                    </div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Compat Data</div>
-                      <div className="crosshook-preview-modal__field-value">
-                        {preview.proton_setup.compat_data_path}
-                      </div>
-                    </div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Steam Client Install</div>
-                      <div className="crosshook-preview-modal__field-value">
-                        {preview.proton_setup.steam_client_install_path}
-                      </div>
-                    </div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Proton Executable</div>
-                      <div className="crosshook-preview-modal__field-value">
-                        {preview.proton_setup.proton_executable}
-                      </div>
-                    </div>
-                  </div>
-                ) : (
-                  <p className="crosshook-preview-modal__empty">No Proton setup data available.</p>
-                )}
-                {preview.trainer ? (
-                  <div className="crosshook-preview-modal__field-list" style={{ marginTop: 16 }}>
-                    <div className="crosshook-preview-modal__sub-label">Trainer Info</div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Trainer Path</div>
-                      <div className="crosshook-preview-modal__field-value">{preview.trainer.path}</div>
-                    </div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Host Path</div>
-                      <div className="crosshook-preview-modal__field-value">{preview.trainer.host_path}</div>
-                    </div>
-                    <div className="crosshook-preview-modal__field">
-                      <div className="crosshook-preview-modal__field-label">Loading Mode</div>
-                      <div className="crosshook-preview-modal__field-value">{preview.trainer.loading_mode}</div>
-                    </div>
-                    {preview.trainer.staged_path ? (
-                      <div className="crosshook-preview-modal__field">
-                        <div className="crosshook-preview-modal__field-label">Staged Path</div>
-                        <div className="crosshook-preview-modal__field-value">{preview.trainer.staged_path}</div>
-                      </div>
-                    ) : null}
-                  </div>
-                ) : null}
-              </CollapsibleSection>
-            ) : null}
-          </div>
-        </div>
-
-        {/* Footer */}
-        <footer className="crosshook-modal__footer">
-          <span
-            className={['crosshook-preview-modal__timestamp', stale ? 'crosshook-preview-modal__timestamp--stale' : '']
-              .filter(Boolean)
-              .join(' ')}
-          >
-            Generated {generatedTimeLabel}
-            {stale ? ' (stale)' : ''}
-          </span>
-          <div className="crosshook-modal__footer-actions">
-            <button type="button" className="crosshook-button crosshook-button--ghost" onClick={handleCopy}>
-              {copyLabel}
-            </button>
-            <button type="button" className="crosshook-button" onClick={onLaunch} disabled={!previewIsReady}>
-              Launch Now
-            </button>
-            <button type="button" className="crosshook-button crosshook-button--ghost" onClick={onClose}>
-              Close
-            </button>
-          </div>
-        </footer>
-      </div>
-    </div>,
-    portalHostRef.current
-  );
-}
-
-/* ───────── LaunchPanel component ───────── */
-
-interface LaunchPanelProps {
-  profileId: string;
-  method: Exclude<LaunchMethod, ''>;
-  request: LaunchRequest | null;
-  profile: GameProfile;
-  /** Profile dropdown (placed in the top row next to Launch / Preview / Reset). */
-  profileSelectSlot?: ReactNode;
-  /** @deprecated Use `profileSelectSlot` — kept for call sites not yet migrated. */
-  beforeActions?: ReactNode;
-  /** Slot rendered where the info/status area is (e.g. pinned profiles). */
-  infoSlot?: ReactNode;
-  /** Slot rendered between the controls card and the actions card (e.g. tabbed config panels). */
-  tabsSlot?: ReactNode;
-  /**
-   * Optional pre-launch gate. Called before launchGame/launchTrainer.
-   * Return true to proceed, false to abort (e.g. show a modal first).
-   */
-  onBeforeLaunch?: (action: 'game' | 'trainer') => Promise<boolean>;
-}
-
-function buildGameOnlyRequest(request: LaunchRequest): LaunchRequest {
-  return {
-    ...request,
-    launch_game_only: true,
-    launch_trainer_only: false,
-    preview_target: 'game',
-  };
-}
-
-function buildTrainerOnlyRequest(request: LaunchRequest): LaunchRequest {
-  return {
-    ...request,
-    launch_game_only: false,
-    launch_trainer_only: true,
-    preview_target: 'trainer',
-  };
-}
+export type { LaunchPanelProps } from './launch-panel/types';
 
 export function LaunchPanel({
   profileId,
@@ -690,23 +55,10 @@ export function LaunchPanel({
 
   const metadata = healthByName[profileId]?.metadata ?? null;
   const versionStatus = metadata?.version_status ?? null;
-  const hasVersionMismatch =
-    versionStatus === 'game_updated' || versionStatus === 'trainer_changed' || versionStatus === 'both_changed';
-  const isUpdateInProgress = versionStatus === 'update_in_progress';
 
   async function handleMarkAsVerified() {
     const outcome = await acknowledgeVersionChange(profileId, revalidateSingle);
     presentAcknowledgeVersionChangeOutcome(outcome);
-  }
-
-  function versionMismatchMessage(): string {
-    if (versionStatus === 'both_changed') {
-      return 'Game and trainer have both changed since last successful launch';
-    }
-    if (versionStatus === 'trainer_changed') {
-      return 'Trainer has changed since last successful launch';
-    }
-    return 'Game version has changed since last successful launch';
   }
 
   useEffect(() => {
@@ -716,14 +68,7 @@ export function LaunchPanel({
   }, [preview]);
 
   const isIdle = phase === LaunchPhase.Idle;
-  const previewDisabled = !request || !isIdle || loading;
-  const validationFeedback = feedback?.kind === 'validation' ? feedback.issue : null;
   const diagnosticFeedback = feedback?.kind === 'diagnostic' ? feedback.report : null;
-  const runtimeFeedback = feedback?.kind === 'runtime' ? feedback.message : null;
-  const diagnosticMatches = diagnosticFeedback ? sortPatternMatchesBySeverity(diagnosticFeedback.pattern_matches) : [];
-  const visibleDiagnosticMatches = diagnosticExpanded ? diagnosticMatches : diagnosticMatches.slice(0, 3);
-  const feedbackSeverity = diagnosticFeedback?.severity ?? validationFeedback?.severity ?? 'fatal';
-  const feedbackLabel = feedbackSeverity === 'fatal' ? 'Fatal' : feedbackSeverity === 'warning' ? 'Warning' : 'Info';
   const launchGuidanceText = [statusText, hintText].filter(Boolean).join(' — ');
 
   useEffect(() => {
@@ -785,190 +130,33 @@ export function LaunchPanel({
           {infoSlot}
 
           {feedback ? (
-            <div
-              className="crosshook-launch-panel__feedback"
-              data-kind={feedback.kind}
-              data-severity={feedbackSeverity}
-              role="alert"
-            >
-              {diagnosticFeedback ? (
-                <>
-                  <div className="crosshook-launch-panel__feedback-header">
-                    <span className="crosshook-launch-panel__feedback-badge">{feedbackLabel}</span>
-                    <p className="crosshook-launch-panel__feedback-title">{diagnosticFeedback.summary}</p>
-                  </div>
-                  <p className="crosshook-launch-panel__feedback-help">{diagnosticFeedback.exit_info.description}</p>
-                  {visibleDiagnosticMatches.length > 0 ? (
-                    <ul className="crosshook-launch-panel__feedback-list">
-                      {visibleDiagnosticMatches.map((patternMatch) => (
-                        <li
-                          key={`${diagnosticFeedback.analyzed_at}-${patternMatch.pattern_id}`}
-                          className="crosshook-launch-panel__feedback-item"
-                        >
-                          <div className="crosshook-launch-panel__feedback-header">
-                            <span
-                              className="crosshook-launch-panel__feedback-badge"
-                              data-severity={patternMatch.severity}
-                            >
-                              {patternMatch.severity}
-                            </span>
-                            <p className="crosshook-launch-panel__feedback-title">{patternMatch.summary}</p>
-                          </div>
-                          <p className="crosshook-launch-panel__feedback-help">{patternMatch.suggestion}</p>
-                        </li>
-                      ))}
-                    </ul>
-                  ) : null}
-                  <div className="crosshook-launch-panel__feedback-actions">
-                    {diagnosticMatches.length > 3 || diagnosticFeedback.suggestions.length > 0 ? (
-                      <button
-                        type="button"
-                        className="crosshook-button crosshook-button--secondary crosshook-launch-panel__feedback-action"
-                        onClick={() => setDiagnosticExpanded((current) => !current)}
-                      >
-                        {diagnosticExpanded ? 'Show Less' : 'Show Details'}
-                      </button>
-                    ) : null}
-                    <button
-                      type="button"
-                      className="crosshook-button crosshook-button--secondary crosshook-launch-panel__feedback-action"
-                      onClick={handleCopyDiagnosticReport}
-                    >
-                      {diagnosticCopyLabel}
-                    </button>
-                  </div>
-                  {diagnosticExpanded ? (
-                    <div className="crosshook-launch-panel__feedback-details">
-                      <p className="crosshook-launch-panel__feedback-help">
-                        Exit mode: {diagnosticFeedback.exit_info.failure_mode}
-                      </p>
-                      <p className="crosshook-launch-panel__feedback-help">
-                        Exit code: {diagnosticFeedback.exit_info.code ?? 'n/a'} | Signal:{' '}
-                        {diagnosticFeedback.exit_info.signal ?? 'n/a'}
-                      </p>
-                      {diagnosticFeedback.log_tail_path ? (
-                        <p className="crosshook-launch-panel__feedback-help">
-                          Log tail: {diagnosticFeedback.log_tail_path}
-                        </p>
-                      ) : null}
-                      {diagnosticFeedback.suggestions.length > 0 ? (
-                        <ul className="crosshook-launch-panel__feedback-list">
-                          {diagnosticFeedback.suggestions.map((suggestion, index) => (
-                            <li
-                              // biome-ignore lint/suspicious/noArrayIndexKey: tiebreaker when severity+title may collide
-                              key={`${suggestion.severity}-${suggestion.title}-${index}`}
-                              className="crosshook-launch-panel__feedback-item"
-                            >
-                              <div className="crosshook-launch-panel__feedback-header">
-                                <span
-                                  className="crosshook-launch-panel__feedback-badge"
-                                  data-severity={suggestion.severity}
-                                >
-                                  {suggestion.severity}
-                                </span>
-                                <p className="crosshook-launch-panel__feedback-title">{suggestion.title}</p>
-                              </div>
-                              <p className="crosshook-launch-panel__feedback-help">{suggestion.description}</p>
-                            </li>
-                          ))}
-                        </ul>
-                      ) : null}
-                    </div>
-                  ) : null}
-                </>
-              ) : validationFeedback ? (
-                <>
-                  <div className="crosshook-launch-panel__feedback-header">
-                    <span className="crosshook-launch-panel__feedback-badge">{feedbackLabel}</span>
-                    <p className="crosshook-launch-panel__feedback-title">{validationFeedback.message}</p>
-                  </div>
-                  <p className="crosshook-launch-panel__feedback-help">{validationFeedback.help}</p>
-                </>
-              ) : (
-                <p className="crosshook-launch-panel__feedback-title">{runtimeFeedback}</p>
-              )}
-            </div>
+            <LaunchPanelFeedback
+              feedback={feedback}
+              diagnosticExpanded={diagnosticExpanded}
+              setDiagnosticExpanded={setDiagnosticExpanded}
+              diagnosticCopyLabel={diagnosticCopyLabel}
+              onCopyDiagnosticReport={handleCopyDiagnosticReport}
+            />
           ) : null}
 
-          <div className="crosshook-launch-panel__profile-row crosshook-launch-panel__profile-row--wrap-centered">
-            <label
-              id="launch-active-profile-label"
-              className="crosshook-label"
-              htmlFor="launch-profile-selector"
-              style={{ margin: 0, whiteSpace: 'nowrap' }}
-            >
-              Active Profile
-            </label>
-            <div className="crosshook-launch-panel__profile-row-select">{profileSelect}</div>
-            <div className="crosshook-launch-panel__profile-row-actions">
-              <button
-                type="button"
-                className="crosshook-button crosshook-launch-panel__action"
-                style={LAUNCH_PANEL_ACTION_BUTTON_STYLE}
-                onClick={() => {
-                  void (async () => {
-                    if (onBeforeLaunch) {
-                      const proceed = await onBeforeLaunch('game');
-                      if (!proceed) return;
-                    }
-                    launchGame();
-                  })();
-                }}
-                disabled={!canLaunchGame}
-                aria-describedby={launchGuidanceText ? launchGuidanceId : undefined}
-              >
-                {isGameRunning
-                  ? 'Game Running'
-                  : isBusy && phase === LaunchPhase.GameLaunching
-                    ? 'Launching\u2026'
-                    : 'Launch Game'}
-              </button>
-              <button
-                type="button"
-                className="crosshook-button crosshook-launch-panel__action"
-                style={LAUNCH_PANEL_ACTION_BUTTON_STYLE}
-                onClick={() => {
-                  void (async () => {
-                    if (onBeforeLaunch) {
-                      const proceed = await onBeforeLaunch('trainer');
-                      if (!proceed) return;
-                    }
-                    launchTrainer();
-                  })();
-                }}
-                disabled={!canLaunchTrainer}
-                aria-describedby={launchGuidanceText ? launchGuidanceId : undefined}
-              >
-                {isBusy && phase === LaunchPhase.TrainerLaunching ? 'Launching\u2026' : 'Launch Trainer'}
-              </button>
-              <button
-                type="button"
-                className="crosshook-button crosshook-button--secondary crosshook-launch-panel__action crosshook-launch-panel__action--secondary"
-                style={LAUNCH_PANEL_ACTION_BUTTON_STYLE}
-                onClick={() => request && requestPreview(buildGameOnlyRequest(request))}
-                disabled={previewDisabled}
-              >
-                {loading ? 'Loading Preview\u2026' : 'Preview Game'}
-              </button>
-              <button
-                type="button"
-                className="crosshook-button crosshook-button--secondary crosshook-launch-panel__action crosshook-launch-panel__action--secondary"
-                style={LAUNCH_PANEL_ACTION_BUTTON_STYLE}
-                onClick={() => request && requestPreview(buildTrainerOnlyRequest(request))}
-                disabled={previewDisabled}
-              >
-                {loading ? 'Loading Preview\u2026' : 'Preview Trainer'}
-              </button>
-              <button
-                type="button"
-                className="crosshook-button crosshook-button--secondary crosshook-launch-panel__action crosshook-launch-panel__action--secondary"
-                style={LAUNCH_PANEL_ACTION_BUTTON_STYLE}
-                onClick={reset}
-              >
-                Reset
-              </button>
-            </div>
-          </div>
+          <LaunchPanelControls
+            profileSelect={profileSelect}
+            launchGuidanceId={launchGuidanceId}
+            launchGuidanceText={launchGuidanceText}
+            canLaunchGame={canLaunchGame}
+            canLaunchTrainer={canLaunchTrainer}
+            isBusy={isBusy}
+            isGameRunning={isGameRunning}
+            phase={phase}
+            isIdle={isIdle}
+            loading={loading}
+            request={request}
+            onBeforeLaunch={onBeforeLaunch}
+            launchGame={launchGame}
+            launchTrainer={launchTrainer}
+            requestPreview={requestPreview}
+            reset={reset}
+          />
 
           <div className="crosshook-launch-panel__runner-stack">
             <LaunchPipeline method={method} profile={profile} preview={preview} phase={phase} />
@@ -982,36 +170,11 @@ export function LaunchPanel({
             ) : null}
           </div>
 
-          {hasVersionMismatch ? (
-            <div
-              className="crosshook-launch-panel__feedback"
-              data-kind="version"
-              data-severity="warning"
-              role="alert"
-              aria-live="polite"
-            >
-              <div className="crosshook-launch-panel__feedback-header">
-                <span className="crosshook-launch-panel__feedback-badge">Warning</span>
-                <p className="crosshook-launch-panel__feedback-title">{versionMismatchMessage()}</p>
-              </div>
-              <div className="crosshook-launch-panel__feedback-actions">
-                <button
-                  type="button"
-                  className="crosshook-button crosshook-button--secondary crosshook-launch-panel__feedback-action"
-                  onClick={() => void handleMarkAsVerified()}
-                  disabled={verifyBusy}
-                >
-                  {verifyBusy ? 'Verifying\u2026' : 'Mark as Verified'}
-                </button>
-              </div>
-            </div>
-          ) : isUpdateInProgress ? (
-            <div className="crosshook-launch-panel__feedback" data-kind="version" data-severity="info" role="status">
-              <p className="crosshook-launch-panel__feedback-title">
-                Steam update in progress \u2014 version check skipped
-              </p>
-            </div>
-          ) : null}
+          <LaunchPanelVersionStatus
+            versionStatus={versionStatus ?? null}
+            onAcknowledge={handleMarkAsVerified}
+            busy={verifyBusy}
+          />
 
           {previewError ? (
             <p className="crosshook-preview-modal__preview-error" role="alert">
