+++ src/crosshook-native/src/components/launch-panel/focusTrap.ts
@@ -1,0 +1,23 @@
+/** Mirrors focus-trap helpers from ProfileReviewModal. */
+
+export const FOCUSABLE_SELECTOR = [
+  'a[href]',
+  'button:not([disabled])',
+  'input:not([disabled]):not([type="hidden"])',
+  'select:not([disabled])',
+  'textarea:not([disabled])',
+  '[tabindex]:not([tabindex="-1"])',
+  '[contenteditable="true"]',
+].join(', ');
+
+export function getFocusableElements(container: HTMLElement): HTMLElement[] {
+  return Array.from(container.querySelectorAll<HTMLElement>(FOCUSABLE_SELECTOR)).filter(
+    (el) => !el.hasAttribute('disabled') && el.tabIndex >= 0 && el.getClientRects().length > 0
+  );
+}
+
+export function focusElement(element: HTMLElement | null): boolean {
+  if (!element) return false;
+  element.focus({ preventScroll: true });
+  return document.activeElement === element;
+}
